Refactor gen_struct_info to allow separate compilation. NFC.
As part of the pending musl update we can no longer (or its no longer
desirable) to compile internal musl headers alongside the public
headers.  To avoid this this change has gen_struct_info.py compile
internal headers seperately.

This change adds support for compiling groups of headers and merging
the results into a single json output.

As part of debugging this I added support for compiling every header
file in islation which is slow but uncovered a couple of issuees.
Structs and defines that were listed under the wrong header and also a
header that could not be compiled on its own (SDL_pixels.h).